+++ zonos/backbone/_mamba_ssm.py
@@ -35,7 +35,7 @@ class MambaSSMZonosBackbone(nn.Module):
 
         self.norm_f = nn.LayerNorm(config.d_model, eps=config.norm_epsilon)
 
-    def allocate_inference_cache(self, batch_size: int, max_seqlen: int, dtype: torch.dtype = torch.bfloat16):
+    def allocate_inference_cache(self, batch_size: int, max_seqlen: int, dtype: torch.dtype = torch.float16):
         return {
             i: layer.allocate_inference_cache(batch_size, max_seqlen, dtype=dtype)
             for i, layer in enumerate(self.layers)
